updated sync-database-test
