+++ test/scripts/database/index.js
@@ -3,5 +3,6 @@
 describe('database', () => {
 
   require('./async');
+  require('./sync');
 
 });
