+++ test/scripts/database/sync/extended/contains.js
@@ -1,0 +1,30 @@
+// jshint esversion: 8
+
+const { assert } = require('chai');
+const db = require('../../../../../database');
+
+const dummyAdapter = { load: () => { return {}; }, save: () => {} };
+
+
+
+describe('#contains()', () => {
+
+  it('test basic contains', () => {
+
+    const object = new db.SyncDatabase(dummyAdapter, {test: "test"});
+    assert.strictEqual(object.contains("test"), true);
+    assert.strictEqual(object.contains("test2"), false);
+
+  });
+
+
+  it('test deep contains', () => {
+
+    const object = new db.SyncDatabase(dummyAdapter, {test: {test: "test"}});
+    assert.strictEqual(object.contains("test.test"), true);
+    assert.strictEqual(object.contains("test.test2"), false);
+    assert.strictEqual(object.contains("test2.test"), false);
+
+  });
+  
+});
