+++ test/scripts/database/sync/index.js
@@ -1,0 +1,9 @@
+// jshint esversion: 8
+
+describe('SyncDatabase', () => {
+
+  require('./extended');
+  require('./reloadData');
+  require('./saveData');
+
+});
